requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.1.3
torch==2.5.1
torchvision==0.20.1
opencv-python==4.11.0.86
numpy==1.23.5
pillow==11.1.0
python-dotenv==1.0.0
flask-wtf==1.1.1
email-validator==2.1.1
flask_cors==5.0.0
pandas==1.4.3
requests==2.32.3
setuptools==75.8.2